move strings into expressions
